@@ -11,11 +11,8 @@ version_file_path = os.path.join(
 
 version = {}
 
-try:
-	execfile(version_file_path, version)
-except:
-	with open(version_file_path) as fp:
-		exec(fp.read(), version)
+with open(version_file_path) as fp:
+	exec(fp.read(), version)
 
 with open("README.md") as f:
     long_description = f.read()
@@ -38,7 +35,7 @@ setup(
         "X3DH>=0.5.9,<0.6",
         "cryptography>=2"
     ],
-    python_requires = ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, <4",
+    python_requires = ">=3.4, <4",
     zip_safe = False,
     classifiers = [
         "Development Status :: 4 - Beta",
@@ -50,14 +47,12 @@ setup(
 
         "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
 
-        "Programming Language :: Python :: 2",
-        "Programming Language :: Python :: 2.7",
-
         "Programming Language :: Python :: 3",
         "Programming Language :: Python :: 3.4",
         "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8"
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9"
     ]
 )
